feat: delegate division of responsibility
