--- README.md
@@ -1,4 +1,0 @@
-# Q&A
-## Question and Answer Platform 
-
-- application currently creates the question and stores it in a URL in JavaScript format
