+++ package.json
@@ -4,7 +4,7 @@
   "description": "questions and answer",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon index.js",
+    "start": "nodemon server.js",
     "test": "jest"
   },
   "repository": "git@github.com:AJohnsoon/qa-plataform.git",
